+++ src/plantuml_gui/sequence/routes.py
@@ -25,7 +25,13 @@
 from flask import Blueprint, jsonify, request
 
 from .activation import add_activation
-from .message import add_message, delete_message, edit_message_text, get_message_text
+from .message import (
+    add_message,
+    delete_message,
+    edit_message_text,
+    get_message_positions,
+    get_message_text,
+)
 from .note import add_note, delete_note, edit_note, get_note_text
 from .participant import (
     add_participant,
@@ -70,13 +76,16 @@ def addmessage():
 def addactivation():
     data = request.get_json()
     puml = data["plantuml"]
-    svg = data["svg"]
     participant = data["participant"]
-    start_y = data["startY"]
-    end_y = data["endY"]
+    start_index = data["startMessageIndex"]
+    end_index = data["endMessageIndex"]
     end_type = data["endType"]
     return jsonify(
-        {"plantuml": add_activation(puml, svg, participant, start_y, end_y, end_type)}
+        {
+            "plantuml": add_activation(
+                puml, participant, start_index, end_index, end_type
+            )
+        }
     )
 
 
@@ -142,6 +151,14 @@ def deletemessage():
     svg = data["svg"]
     svgelement = data["svgelement"]
     return jsonify({"plantuml": delete_message(puml, svg, svgelement)})
+
+
+@sequence_bp.route("/getMessagePositions", methods=["POST"])
+def getmessagepositions():
+    data = request.get_json()
+    puml = data["plantuml"]
+    svg = data["svg"]
+    return jsonify({"positions": get_message_positions(puml, svg)})
 
 
 @sequence_bp.route("/addNote", methods=["POST"])